System/website/requirements.txt
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
Babel==2.8.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-fontawesome==1.0
django-phonenumber-field==4.0.0
firebase-admin==4.2.0
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauth2client==3.0.0
phonenumbers==8.12.2
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycryptodome==3.4.3
PyJWT==1.7.1
pylint==2.5.0
PyMySQL==0.9.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
twilio==6.39.0
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
